Lowered the res, increased the awesome
more droplets? smoother fluid? Increased the size by rendering to a
smaller grid and stretching it. so the droplets are smaller and there
are more, but they're being rendered to a smaller area than before and
then that area is being scaled to fullscreen to make it appear the same
size as before (fullscreen, that is).